Updated README.md
In README file, there is a docker run command to create the docker container. 
$ docker run -d -p 8000 mhowlett/ngx-stub-status-prometheus
This command exposes the container port 8000 but it does not bind local port 8000 to container. So when user tries to access http://127.0.0.1:8000/metrics URL, it does not work. As by default docker takes some random local port and binds it to container's 8000 port.